+++ .gitignore
@@ -43,3 +43,4 @@ Thumbs.db
 node_modules/
 bower_components/
 src/vendor/
+coverage/
