--- src/gtten.js
@@ -1,7 +1,0 @@
-module.exports = function(x) {
-    if (x > 10) {
-        return true
-    } else {
-        return false
-    }
-}
